+++ src/main/java/net/lewmc/kryptonite/commands/ExploitDBCommand.java
@@ -1,8 +1,8 @@
 package net.lewmc.kryptonite.commands;
 
 import net.lewmc.kryptonite.Kryptonite;
-import net.lewmc.kryptonite.exploitdatabase.Check;
-import net.lewmc.kryptonite.exploitdatabase.Patch;
+import net.lewmc.kryptonite.edb.Check;
+import net.lewmc.kryptonite.edb.Patch;
 import net.lewmc.kryptonite.utils.MessageUtil;
 import net.lewmc.kryptonite.utils.PermissionUtil;
 import net.lewmc.kryptonite.utils.SoftwareUtil;
