+++ src/main/java/net/lewmc/kryptonite/kos/ServerProperties.java
@@ -1,4 +1,4 @@
-package net.lewmc.kryptonite.optimiser;
+package net.lewmc.kryptonite.kos;
 
 import net.lewmc.kryptonite.utils.PropertiesUtil;
